Extract JS from React lib
